requirements.txt
pydantic
python-dotenv
fastmcp
httpx
aiocache
aiolimiter